requirements.txt
Django==2.2.7
django-cors-headers==3.2.0
django-crispy-forms==1.8.1
djangorestframework==3.10.3
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0